Sequencer: simplify channel override entity import, fix float track channel override support
Before, channel overrides would get a type-erased parameter that gives information about what should be overriden and how. This was flawed, since a given type of channel can't know what kind of parameter was created by the caller, since various types of tracks and sections could have handled that. So we just simplify this for now, forcing the parameters to be (for now) just a channel name and result component. That is: parameters that tell the override which channel it's posing as, and where to put the result.

A second bug was that we would still create an ECS entity even if *all* of a section's channels had been overriden, which is easy to do when overriding the channel of a float track (since it only has 1 channel). This was particularly bad since it would allocate the entity but not initialize it with proper data. Now, the import helper checks whether there are non-overriden channels left before allocating space for an entity.

Finally, the float section was incorrectly giving any override channel a source component instead of a result component. The import helper API was to blame here, so this had to also be fixed.

#rb max.chen
#preflight 6320b9a529254beccb788c8f

[CL 21988206 by ludovic chabant in ue5-main branch]