Improved import setting button behaviour:
When users click on import and open an appearance setting, it will automatically apply the theme if it exists. If not, prompt users to import a theme from the same location as the Appearance setting config file. If import succeeds, the new theme is applied and the .json file is copied to user-level theme directoy. If users choose not to import, keep the current theme.

#jira UE-160593: Update Theme Saving Location
#rb Sebastian.Nordgren
#preflight 631a105fd31788ea3a46ff72

#p4v-cherrypick 21981452

[CL 21989021 by lauren barnes in ue5-main branch]