FORT-514478 - Blueprint struct is not determining whether the floats are single/double precision
We don't need to assert if we fail to find a conversion function for a struct type. In fact, this is a perfectly valid scenario in some cases. For example, execConvertMapType will check if either the key or value has a conversion function. If it doesn't, then it'll just rely on the property's default copy behavior. In many cases, the key/value might be a struct that doesn't have a conversion entry (ie: a struct that isn't a core type, such as FVector).

#preflight 632138e0e93a80888c75b0cd
#rb justin.hare

[CL 22009855 by dave jones2 in ue5-main branch]