Changed the InstancedStaticMeshComponent component creation so that it has hit proxies for its instances by default when it is not dynamically or natively created.
This make the instances selectable in the editor at all time. Before that user needed to add a instance via the details panel to get that behavior. That was confusing and it didn't allow editor script to create those selectable instances.

Also override the OnComponentCreated function for the UGrassInstancedStaticMeshComponent because they don't need per instance hit proxies.

#jira UE-161646
#rb Brooke.Hubert, Jamie.Dale
#preflight 6322185f8c478acecfccf828

[CL 22012196 by Julien StJean in ue5-main branch]