Sequencer: Don't create useless entities when we are dealing with a muted override channel
It's possible for PopulateEvaluationField to create evaluation field entries for muted override channels if it doesn't have access to the editor data that tells us if a channel is enabled or not. In this case, we will be asked to create an entity builder but we will skip most of it since we will realize the channel was muted. There was a bug however where we would still add entity builders for the basic stuff (property binding, object binding, easing, etc). This would of course do nothing useful and was just wasteful. So this fix only adds these entity builders once we know the channel is good to go.

#jira none
#preflight 6334e2e9f418a1071a3b6da7
#rb andrew.rodham, max.chen

[CL 22263140 by ludovic chabant in ue5-main branch]