Fixed missing trace events.
Trace uses a thread local object and it destructor as a means to collect buffers when threads end. This usually happens lazily, and usually when fetching a new identified for a thread. On some platforms can result in an allocation and if memory tracing was on, its events would end up in the redirect buffer as the correct buffer hadn't yet been set. But because the redirect had already been memcpy'd, events would get lost.

#rb im
#rnx
#jira UE-157559
#preflight 633a7e975c2225fe5f32a5f0

[CL 22295530 by martin ridgers in ue5-main branch]