UGS: Allow overriding the filter used to determine if a file contains code.
The CodeFilter property is read from the [Perforce] section in the config file, and is appended to the default list of filters used to find code changes (having a cpp/h extension and so on). Syntax is in standard Perforce syntax, with a '-' prefix to exclude matching files. For example, the following entry..

  [Perforce]
  CodeFilter=-.../Experimental/...
  CodeFilter=-*.h

Excludes all *.h and files under an /Experimental/ folder anywhere in the tree from being treated as code changes.

#fyi Mike.Beach
#preflight none

[CL 22541814 by Ben Marsh in ue5-main branch]