Fixed a prediction issue when using AbilityTask_NetworkSyncPoint. Abilities executing gameplay cues on a local predicted ability were not ever playing the gameplay cue locally.
Repro:
1) Create a burst gameplay cue that does some simple particle effect and gameplay cue tag to associate with it
2) Create a Local Predicted/Instanced Per Actor GA that when it activates will start a Delay Node (2s was my test) then does a Wait Net Sync (Wait Both) after the sync call ExecuteGameplayCueOnOwner for the gameplay cue tag
3) Grant then activate the GA on the client

If running multiple clients all simulated clients see the burst cue fine, but the local client won't

Reason and then fix:
In this code the ScopePrediction was being made properly and passing from client to server so all predicted paths properly ignored execution assuming the client will/already handle those executions. But, because of this node being "latent" as soon as it finishes executing waiting for the delegate to fire it lost the prediction key made during local activation, and thus wouldn't perform any local predicted executions. By moving the prediction key to only when the server event fires and then just before broadcasting the delegate for the action to resume the ability we properly setup a ScopedPredictionKey allowing local clients to properly run predicted executions.

#abilitysystem
#rnx
[REVIEW]

[CL 22573156 by todd eckert in ue5-main branch]