UE-112211 - Crash in UKismetSystemLibrary::execSetStructurePropertyByName
Getting a reference to an element via UKismetArrayLibrary::Array_Get in an array of user structs can lead to a null source pointer in Generic_SetStructurePropertyByName. Specifically, this happens if the array is empty, or an invalid index is used.

We correctly throw a BP exception when this scenario occurs. However, execArrayGetByRef will set Stack.MostRecentPropertyAddress to null, which is then used as the source parameter to Generic_SetStructurePropertyByName. Returning a copy of the element doesn't have the same issue since it creates a temporary user struct on the stack as a source struct.

The most straightforward fix is to simply soft-fail in Generic_SetStructurePropertyByName. The user will have to rely on the BP exception info for the actual error.

#jira UE-112211
#preflight 634dd931d737d61a2f156c8c
#rb phillip.kavan

[CL 22605284 by dave jones2 in ue5-main branch]