Chaos Cloth: Add damping to XPBDSpring and XPBDBending constraints.
Refactored PBDStiffness so that it has a new base class PBDWeightMap which can be used for the damping maps. The damping maps do not do the exponential conversion that PBDStiffness::ApplyValue has.

XPBD constraints are not available through the UI, so no change in behavior for users. As usual, the damping parameters exist at the ChaosClothingSimulationCloth and Constraints level, but they are not exposed to the Config, so they are defaulted to 0. Similarly, the map values exist at the Constraint level, but they are not populated in the UI Weight targets.

For both constraints, added Rayleigh damping following the methods described for XPBD. The exposed damping parameter is the DampingRatio (where 1 = critical damping). For efficiency, there are separate ISPC kernels for when DampingRatio > 0 vs no damping since additional computation is necessary for damping.

#preflight 6350734f3833996f641d1164
#rb kriss.gossart

[CL 22677669 by Alex McAdams in ue5-main branch]