VR Editor: Introduce minimal UVREditorModeBase class.
This is the new parent class of UVREditorMode, and subsumes a small subset of its functionality (currently just the HMD stereo viewport rendering setup). IVREditorModule and FVREditorModeManager now use pointers to the base class type. Some deprecation annotations were added, but are commented out and will be enabled in a subsequent changelist due to the large number of external files that require accompanying fixups.

Other changes:
 - The analytics events have been moved from UVREditorMode into FVREditorModeManager (we always want these).
 - The use of UViewportWorldInteraction has been moved from FVREditorModeManager into UVREditorMode (this was really an implementation detail of UVREditorMode).
 - Support for the "Entering VR Mode - Experimental" warning, which was only used by UVREditorMode, was removed.
 - The UVREditorModeBase now has the UWorld as an outer, which makes it eligible for selection in the blueprint debugger.

#jira UE-167701
#rb jason.walter
#preflight 6351cc170313c24974e8f3ad

[CL 22695140 by zach brockway in ue5-main branch]