Hololens + D3D12: Disable GPU crash tracking data
D3D12 uses WriteBufferImmediate to write breadcrumbs for GPU crash tracking. Hololens claims support ID3D12GraphicsCommandList2, but inexplicably, does not support WriteBufferImmediate (the only additional method exposed for GraphicsCommandList2). This is probably because the underlying QCOM implementation does not support the command.

We just disable the entire GPU crash tracking setup on Hololens, as it is dependent on the breadcrumbs.

#jira UE-167021
#preflight 635249cc7261e565c45dfa50
#rb Arciel.Rekman

[CL 22712204 by robert srinivasiah in ue5-main branch]