Introduced a new plugin to host the Editor Framework's new data storage.
This changelist introduces a new plugin that provides an implementation for the Editor Framework's new data storage. The data storage can be seen as a central database-like structure that eventually will become the single source of truth for all data used by the editor. Processors can be added that (periodically) run over query results allowing arbitrary data to be displayed that matches the data. This in turn can be used to create reusable UI elements and other functionality. Users can add their own columns with data for custom and novel functionality as well.

This first addition introduces:
- The plugin itself.
- Two new interfaces to interact with the data storage. To be expanded upon in the future.
- Updates to the Typed Elements API to access the new API.
- Automatically start using the new data storage if the plugin is enabled.
- Register any actors in the editor and unregister when they get garbage collected or explicitly destroyed.

Other than being able to see the number of actors in the MASS Debugger, there's no no or replacement functionality introduced.

#rb brooke.hubert
#preflight 6356d407777a77c4404435af

[CL 22733276 by ronald koppers in ue5-main branch]