Add custom version numbers for PCG. These open the door to additional powerful migration use cases in addition to the usual methods such as the '_DEPRECATED' postfix.
The initial use case of this will be projection node inputs migration - a single input is now split into projection source and target. It was not trivial to detect the split inputs from the deprecation code.

This kind of migration of node connections must happen on/after graph deserialization rather than node (settings) deserialization, because it is a graph-level fixup. And also the node pins/edges are not resolved until after Serialize(), so I could not simply apply fixup in Serialize() as would normally be the case. So this particular migration is not a good example of using the custom version, and is more complicated than normal migration. However the custom version implementation is powerful and is hopefully usable in an easier way for other migration needs in the future. The custom version can be used on any UObject in the PCG system using standard patterns used throughout the engine.

I have added an additional custom version that we will not touch internally, which external teams could use if they wish.

#rb julien.lheureux
#preflight 635796f79e14ee3c798de8c1

[CL 22795959 by huw bowles in ue5-main branch]