UE-162245 - Level asset doesn't duplicate correctly with Event Dispatcher in Level BP
When duplicating a level with an event dispatcher, the signature function was actually pointing to the original UFunction that belonged to the UClass in the level that we were copying from. As a result, a duplicated level would have object references outside of its current package, which prevented the editor from saving it.

The fix updates ClassGeneratedBy to point to the duplicated ULevelScriptBlueprint. Otherwise, FMemberReference might resolve to the class/function that lives in the source level.  Also note that standard duplication will not perform a deep copy of this field since it's outered to a different object. Similarly, we need to regenerate the CDO for our new class to prevent any save errors.

#jira UE-162245
#preflight 635af38fab23457f85bc8363
#rb dan.oconnor,phillip.kavan

[CL 22833317 by dave jones2 in ue5-main branch]