Hide unused payload structs from shaders which did not declare them
This is meant to help avoid inconsistencies between GetRayTracingPayloadType() and the shader code.

As an example of the usage, convert RayTracingDebug related shaders to only enable the payload conditionally. Fix the current incorrect mixing of payload uses in the same shader by introducing a permutation for the material version of the debug shader vs. the one that uses the debug payload (this was previously a runtime shader parameter only which implied two different payloads could theoretically be compiled into one RTPSO).

Also conditionaly enable some of the simpler payload types like the niagara and decal ones that are only used in a few places. Other payloads will be handled in future refactors.

Split RayTracingDebug.usf into smaller files that only have one raytracing entry in them to reduce the amount of counditional logic needed around payloads.

#rb Yuriy.ODonnell
#jira none
#preflight 635c49e9ae6840072d4df82f

[CL 22849513 by chris kulla in ue5-main branch]