Fixed UGatherTextFromAssetsCommandlet failing to gather text from external actor packages
UGatherTextFromAssetsCommandlet wasn't finding external actor packages for the worlds it processed, nor was it correctly loading those packages even if it did find them.

UGatherTextFromAssetsCommandlet has been updated to do a secondary scan for external actor packages that belong to any world partitioned worlds that it is going to gather localization for. Ideally these external actor packages will be gathered using their cache, but their caches are currently invalid due to the FPropertyLocalizationDataGatherer issue.

In the cases where we detect a stale external actor localization cache, we instead queue its owner world to be re-gathered with the GUID(s) of any external actors that were stale. This takes the world package through a different path in the package gather code that will load and initialize the editor world within the package, and use the world partition utils to iterate and load any of the stale actors so that they're in a valid state to be gathered from.

Note: The version bump to use the localization cache for external actors that have been re-saved after the FPropertyLocalizationDataGatherer fixes will be made when this change reaches Main.

#jira
#preflight 63612013e271e75dea8c13e2
#rb Rex.Hill, JeanFrancois.Dube

[CL 22890158 by jamie dale in ue5-main branch]