Sequencer: Fix shot parsing so that if the current prefix is found at the beginning of the shot name, it will be stripped and the parsing of the shot number and take number will continue from there.
This fixes a situation where shot prefixes can collide with parsing shot and take numbers because they are looking for the same separators. For example if the shot name is: sq050_shot0010_01, that should be parsed as shot 10 and take number 1. If the user specifies that the shot prefix is sq050, that portion can be stripped so it doesn't erroneously parse as shot 50.

#jira UE-168530
#preflight 63606507ce68f7cbb6560cdc
#rb matt.hoffman

[CL 23104560 by max chen in ue5-main branch]