Integration of CL: 23064620 #jira: FORT-537062
Change when ImageNtHeader is called to PreInitPostStartupScreen. This is after some hidden DbgHelp initialization is done so future calls to SymFromAddr succeed. There is not a clear answer for when ImageNtHeader can be called but where it was before was too early.
This is potentially an issue with large PDBs
This issue only was seen to reproduce in monolithic exes

[REVIEW] [at]Bob.Tellez [at]Devin.Doucette [at]Robert.Millar [at]Gary.Yuan [at]Joe.Kirchoff
[REVIEW] [at]Devin.Doucette [at]Robert.Millar
#preflight 636c0f617c2b505190240d9b

[CL 23119748 by eric knapik in ue5-main branch]