Sequencer: Fixed crash editing object channel keys
The actual crash was caused by a disparity between the actual channel value type (FMovieSceneObjectPathChannelKeyValue) and the type used to edit the value (FSoftObjectPtr).
When constructing an instance of the key struct, we would write into the instance property as if it were a FMovieSceneObjectPathChannelKeyValue, and trample memory. Ordinarily this doesn't happen because the types match, but object channels override the types here.
To fix this I made it the responsibility of PostConstructKeyInstance to assign the value, allowing the object path channel to correctly populate the value.

I also found that the CPP struct ops were not working correctly (something must have changed since this was first written; this class has a virtual destructor but its vptr was always nullptr because the struct ops were not being set-up/used any more), and in fact it is not well supported to have both CPP struct ops _and_ non C++ defined UPROPERTIES on a UScriptStruct. To this end I have removed the cpp struct ops from the generated struct and added code to do the initialization and destruction of the struct instance directly.

#jira UE-169382
#preflight 6373f23d0c74adb48ba56acc

[CL 23141301 by Andrew Rodham in ue5-main branch]