PatternTool: Use a gizmo to allow users to modify extents/radius in the viewport visually.
A single gizmo whose position relative to the tool's construction plane mechanic's gizmo is used to define Extent/ExtentX & ExtentY/Radius. The gizmo's movement is restricted depending on the current shape, axis/plane, and the orientation of the tool's construction plane frame. The visual appearance of the gizmo is modified by subclassing the actor builder for the existing combined transform gizmo and registering a 'new' gizmo type with the gizmo manager which is based on the existing combined transform gizmo.

#jira none
#rb semion.piskarev
#preflight 6373f666b663683828d066fd

[CL 23155451 by Nickolas Drake in ue5-main branch]