The Typed Element queries have been updated with support for a count query and the initial step to run queries.
Count queries are queries that don't select any columns but instead return a count of the total number of rows that pass the filter conditions. A call was also added that allows these queries to be run, which is currently limited to retrieving the result of a count query as the other query type, select, can't be direclty run.
This change list also introduces FTypedElementDataStorageColumn and FTypedElementDataStorageTag as specific base classes for columns tags. These are used in upcoming changes.

#preflight 63752b1fbf76990b71d527d3
#rb brooke.hubert zach.rammell jeanmichel.dignard

[CL 23163224 by ronald koppers in ue5-main branch]