Clear keyboard focus when setting an empty set of selected objects on an SDetailsView. This commits any pending changes before the widgets are destroyed.
This is necessary because although we register an OnRowReleased handler, we also register a Visibility handler in SDetailsViewBase::GetTreeVisibility(). The OnRowReleased delegate is called when the tree next Ticks(), but when the tree is cleared, it also gets its visibility set to Collapsed, which causes the next Tick() to not occur. Since there's no way to force the tree to Tick(), this is the next best thing.

Added a few trace macros.

#jira UE-162823
[REVIEW] [at]patrick.boutot
#rnx
#preflight 6376176b4f460e024c15a9cf

[CL 23174550 by sebastian nordgren in ue5-main branch]