[UE-168454] Some Watch Pins Show Incorrect Values in the Player AnimBP
The source of the issue was that the key of the mapping between the original and the folded property was not able to distinguish between the instance of the node. This means that e.g. when we had several blend nodes in the graph, only one of the properties was in the map and thus all pin watches were watching the single folded property that the map contains.

This is now solved by using a graph pin as key for the map, which is unique per node instance.

An additional issue was that nodes like get variable were not compiled by the blueprint compiler and thus their pins were not part of the map as well. This got fixed by adding linked pins to the map as well.

The fix got verified in game as well. Please make sure to reload/recompile the "avg_player" and "sourcepose" anim bp before verifying.

#jira https://jira.it.epicgames.com/browse/UE-168454
#preflight 637e36baf514e1ded9722abe

[CL 23249548 by benjamin jillich in ue5-main branch]