[PCG] Rework projection logic, fix projection of landscape splines, and support projection params on spline projection.
We were forcing collapse to points when sampling general (non-spline) projections, because projection is not a 1:1 function (many points in source can project to same point on target). However this is not the case if ProjectionParams.bProjectionPositions==false because nothing actually moves, in which case we can stick to analytic sampling. Therefore RequireCollapseToSample() uses this as the criteria.

The spline projection case with ProjectionParams.bProjectionPositions==false falls back to the base projection code, which will do a tidy sampling of the 3D spline (as we are not moving the spline positions - quirky but is consistent with processing of 'concrete' data (points)).

The landscape spline makes use of a similar path. It is probably possible to avoid this new UPCGLandscapeSplineProjectionData class completely and put logic for detecting this case in UPCGProjectionData::RequiresCollapseToSample().

Rename projection params to disambiguate.

#preflight 63865238d76995cb1455910a
#rb julien.lheureux, adrien.logut

[CL 23316957 by huw bowles in ue5-main branch]