Introduces a new world type "EditorStorage".
The "EditorStorage" world is meant to be used to store data for the entire lifetime of the editor process. As such it's not destroyed when new levels are loaded and is meant to be constructed with minimal setup to avoid overhead. This new world type is currently used in the Typed Elements Data Storage which serves as the central database-like storage for the editor and is powered by MASS. The latter currently requires a world for interacting with the outside through subsystems and to be periodically ticked to make sure data remains consistent. The changes that make use of this will be submitted separately to keep this CL focused.

#rb Ben.Zeigler
#jira UE-171267
#preflight 638a816535192facc1dffbb5

[CL 23396289 by ronald koppers in ue5-main branch]