USD: Fix handling of published assets when importing with asset replace policy
This finally fixes that annoying ensure on that line "ensureMsgf(false, TEXT("Texture not valid! UE-23902! Parameter (%s)"), *Parameter.ParameterInfo.Name.ToString());". You can see on the Jira that I was just working on 5.2 for this, but noticed this is happening on 5.1 as well. Not sure why I got a Jira about the 5.2 issue and nothing about the 5.1 issue?

You can see the same description on the comment, but basically there were two problems here:

The first problem is that when we import assets with EReplaceAssetPolicy::Replace, *and* we run into already existing UAssets during PublishAsset, we'll end up using DuplicateObject to "stomp" the newly imported asset on top of the existing asset (that is, we'll do the "Replace"). The actual issue here is that the asset that will remain on the "UsedAssetsAndDependencies" set (which we continue reusing throughout the import) is the original, transient asset! On the other EReplaceAssetPolicy options (or even on EReplaceAssetPolicy::Replace if we don't find an existing asset to stomp), we'll actually move the same imported asset onto its published package, so while the same asset remains on UsedAssetsAndDependencies, it goes from the transient package to the published package. When we end up using DuplicateObject though, the "new" duplicated object doesn't go into the UsedAssetsAndDependencies, but the original asset remains there, even though it remained on the transient package.

This first problem is fixed by using GetPublishedAssetsAndDependencies after doing all the replacements, and using the set of published assets from that point onwards (which is what the code was expecting in the first place).

The second problem is that apparently material instance proxies have direct pointers to the UTexture assets that they're referencing, *separate* from the UMaterialInstance's pointers. When we call our ReplaceReferences and ReflaceSoftReferences after publishing stuff, we remap the UMaterialInstance's references to the published textures, but we can't reach into the material proxies to do the same. You can see how this becomes an issue when it is combined with the first problem: It will then be possible to get these proxies pointing at transient textures, that can eventually get fully collected. Once the engine allocates a new UObject onto wherever address the proxies are pointing, some code realizes its texture parameters are not pointing at textures anymore and freaks out (you can see on Horde error message above the callstack how it prints that it's texture pointer was pointing at ShinyBlue_327 instead, which is a material).

The second problem is fixed by calling PostEditChange on the imported materials. Luckily that is enough to get it to regenerate the material proxy, which recreates its texture pointers from the material instance's, which do get remapped successfully.

#jira UE-171227
#rb anousack.kitisa
#lockdown jeanmichel.dignard
#preflight 638daef335192facc16db1b3

[CL 23421882 by daniel coelho in ue5-main branch]