Fix sparse class data crash in the presence of circular dependencies
This ensures that 'old' sparse class data (that may be regenerated by a compilation) is always stashed when beginning a compilation pass. This is needed because when sparse class data is cleared, all (loaded) child classes are also cleared, so when they also come to be compiled, their sparse class data will already be null and there will be a dangling reference to it in the package's export table (despite it being moved into the transient package). The dangling reference will then cause any subsequent loads (e.g. further children) to reference an invalid structure.

#jira UE-171639
#rb Jurre.daBaare
#preflight 63906af667018b14b5fe96be

[CL 23431859 by thomas sarkanen in ue5-main branch]