Chaos Cloth: minor optimization to graph coloring to take the ParticleOffset, which is the start particle index for that cloth. This is an optional parameter. Default behavior is the same as existing behavior.
ParticleOffset should be <= the first node in the graph. This will reduce graph coloring particularly if the ParticleOffset is very large by reducing local allocations and initialization.

Updated the cloth constraints to pass their ParticleOffset to the graph coloring. Did not update the volumetric/flesh constraints because they do not know ParticleOffset already (they don't have weight maps, which is why the ParticleOffset is passed to constraints currently).

Tested with Venus Pantsuit which has two cloths and verified (in debugger + running with Debug checkSlow enabled) that everything is working as expected.

#rb kriss.gossart
#preflight 6390e2af0d013d47ef41c6ee

[CL 23436500 by Alex McAdams in ue5-main branch]