Fix crash when resizing the window.
FD3D12Queue::ExecutePayload does not increment the fence value if there hasn't been any useful work (command lists to submit) since the last invocation. However, this function is not aware of Present calls, since they happen outside of the submission pipeline. This means that the empty payload submitted by BlockUntilIdle can fail to insert a fence wait if a present happened after the previous payload was executed. This reproduces very easily when resizing the window, but could happen during normal rendering as well. The short-term solution is to force the BlockUntilIdle sync payload to always signal the fence. The correct solution is to execute Present calls as part of the submission pipeline, but this is a more involved change which will happen at a later date.

Also fix numeric error codes in GPU crash messages, instead of the corresponding DXGI_ERROR_* name. This happened because the submission code used VERIFYD3D12RESULT() directly on the error code returned by GetDeviceRemovedReason(), which is incorrect. We must pass the device removed code to VerifyD3D12Result when we detect the error, and let it call GetDeviceRemovedReason to print the correct reason code.

#rnx
#rb Kenzo.TerElst
#preflight https://horde.devtools.epicgames.com/job/638f5dfcc0652bbec25b5da4

[CL 23445818 by mihnea balta in ue5-main branch]