Lumen - replace cone skylight evaluation with a ray sample in SWRT radiosity.
There's no point in having cone evaluation only for a single place in the pipeline. Additionally this was causing a lot of leaking with bright skylights. E.g. when a ray hits black lower sky hemisphere it still could bring in color from the upper one due to the ray footprint.

#fyi Daniel.Wright
#preflight skip

[CL 23446595 by Krzysztof Narkowicz in ue5-main branch]