Sequencer: Fix level sequence project settings not persisting because they are tied to cvars.
The bug is that when you reload the engine, the config is properly restored, but are then overwritten by the corresponding cvar value. That means if you edit the setting through the UI or set it through ini, the default value from the cvar will win out. This is not desirable.

Tested the following:
- If you edit the project setting, restart, and the project setting will persist. A created level sequence will have that same setting.
- If you edit the ini, restart, and the project setting will persist. A created level sequence will have that same setting.
- If you change the cvar (either on load or after the engine has loaded), the project setting will not be updated. Any created level sequence during this session will have that setting.

#jira UE-171985
#rb matt.hoffman
#lockdown laurent.delayen
#preflight 63912e9dbb6fefa4724aba49

[CL 23460502 by max chen in ue5-main branch]