[PCG] Second iteration of disabled node behaviour.
Firstly tweak pass through behaviour - check types overlap and the input type(s) do not exceed what the output type supports. Some types are only known dynamically/at execution time (there are currently nodes that take Any and output points, for example, and we don't know what data will come in before execution), so type checking happens during pass through. An example of what this solves: a node that takes a surface as input and outputs points. This should *not* pass through a surface when the node is disabled, as it was supposed to produce points. Its job was to convert a surface to points, if it is disabled, it outputs 0 points.

Secondly when node is disabled, pass through all edges connected to pass-through pin by default. Basically assume node operation is "for each edge, ...". The only nodes that actually perform an operation across all edges are Union, Intersection and Merge and these will pass first edge only when disabled.

To support precise type checking, UPCGData::GetDataType narrowed to give most specific type.

#rb julien.lheureux
#preflight 6396ff780056adf2560d1404

[CL 23499848 by huw bowles in ue5-main branch]