[PCG] Change SelfPruning input/output types from Any to Point.
This node converts to points internally. In the future it may deal with more than just points, if so we can widen the types accordingly.

Add Param pin to catch first existing params connection (if any), and add deprecation to automatically set it up to not break existing graphs. Note that as per noding standard Params pin does not support multiple connections and only takes the first.

To support this, multiple changes made to node deprecation:

* Logic made less flimsy by asserting DataVersion is up to date after deprecation
* New deprecation hook added that is called before UpdatePins. This gives the node a chance to rearrange connections before UpdatePins removes any invalid connections. We can use this to find any connection to Params and move it to a params pin. (My other local change that splits the Surface Sampler pins also requires this deprecation hook).

#rb julien.lheureux
#preflight 63974ac98c64c74ac81e427a

[CL 23499884 by huw bowles in ue5-main branch]