[PCG] TransformPoints: Apply concrete, narrow types (types that are used internally in the node). This does not kick off any connections in existing graphs due to our lenient type checking.
If we extend this node to process more than just points in the future we can widen the types accordingly.

Also reorder pins so Params is second. Interestingly this does nothing to existing node usages in the graph, but will apply to new ones. Potentially UpdatePins() could be extended to ensure pin orders are correct, but having it apply to new node instances while preserving existing did not seem like the worst behaviour to me.

#preflight 63974a25c16855964d3f81d3
#rb julien.lheureux

[CL 23499885 by huw bowles in ue5-main branch]