ModelingMode: Add support for Gizmo snapping and pivot-repositioning in new Selection mode.
UDragAlignmentMechanic is refactored into FDragAlignmentBase with subclasses UDragAlignmentMechanic and UDragAlignmentInteraction, the latter is usable outside of a Tool (the Mechanic requires a Tool).
FFindSceneSnapPointParams struct in ToolSceneQueriesUtil now takes explicit SnappingManager and CameraState, instead of a Tool that is used to look up those values. Update the one call site (inside of ToolSceneQueriesUtil) that used this struct.
Add a DragAlignmentInteraction to ModelingSelectionInteraction and hook up to TransformGizmo, as well as required DragAlignmentToggleBehavior. Also required adding UModelingSelectionInteraction::Render(), and calling fro ModelingToolsEditorMode

#rb none
#preflight 639b95f72540a78d27047838

[CL 23535316 by ryan schmidt in ue5-main branch]