Path Tracer: Apply roughness clamping and shading normal adjustments in the material instead of the integrator
This will be required for Strata as we won't be able to modify the encoded BSDF tree after it has been returned by the material.

Refactor how optional payload fields are managed as a consequence of this re-organization. The SSS profile is now partially decoded inside the material. This reduces the number of shading model specific operations that need to happen inside the integrator. More work will shift to the material in future refactors.

There are now 8 custom data fields in the payload, used by the various shading models to encode additional information.

#rb none
#jira none
#preflight 63b34a43776b61ba3bff5cd6

[CL 23563620 by chris kulla in ue5-main branch]