Avoid clearing the undo buffer when saving newly added/deleted OFPA actors
A newly created OFPA actor would trigger PostLoadedStateChanged to clear the undo buffer due to it creating a new actor desc. This now skips clearing the undo buffer unless an actor has been unloaded.

A deleted OFPA actor would trigger CleanupAfterSuccessfulDelete to clear the undo buffer so that the package can be GC'd. This makes sense for assets (where a delete isn't transacted), but not for actors (where the delete can be undone). This now skips clearing the undo buffer when cleaning-up OFPA packages, and also marks any lingering packages as being PKG_NewlyCreated (since they no longer have a file on disk) and dirty so that they can be saved again if the delete is undone.

#preflight 63b449f435203bc7aaa75fa2
#rb Francis.Hurteau

[CL 23570251 by jamie dale in ue5-main branch]