nDisplay[SharedMemoryMedia]: Added media source modes of operation for non nDisplay usage.
Framelocked - Matches source and local frame numbers. Always use this mode in nDisplay.
Genlocked - It doesn't match frame numbers, but it also doesn't skip frames, so will hold back the sender (up to a timeout) if it is faster than the receiver.
Freerun - It always grabs the latest frame. It may skip frames if they arrive too fast.

#jira
#rb andrey.yamashev
#preflight 63c6275ba06ab115eaf3f13a

[CL 23735129 by Alejandro Arango in ue5-main branch]