DebugSkelMeshComponent: Do not clear the preview instance (or mark it for garbage collection) when clearing anim script instance.
This is a followup fix to CL 23407330. The original bug was caused by calling ClearAnimScriptInstance (which would mark the AnimScriptInstance for garbage collection) when the PreviewInstance is pointing to it. The original fix was to set the PreviewInstance to nullptr after this happened. However, this broke a few other things, namely RBAN simulation and PhysicsAssetEditor preview animation stopped working because the PreviewInstance was expected to exist.

Upon review, thomas.sarkanen suggested that the MarkAsGarbage was completely unnecessary, so just not doing that.

Tested original bug  (UE-171723) was still fixed, PhysicsAssetEditor preview animation works, and RBAN preview simulation still works.

#preflight 639bbc7943330e63e5943c0c
#rb thomas.sarkanen helge.mathee

[CL 23759855 by Alex McAdams in ue5-main branch]